# botscreen-public/server/openapi.yaml
openapi: 3.0.3
info:
  title: 互动问答后端 API
  description: |
    botscreen 互动问答与语音唤醒后端接口清单。
    本清单依据 `server/qa_server.py` 的实际 FastAPI 路由整理。
  version: 3.0.0
servers:
  - url: http://127.0.0.1:8000
    description: 本机问答服务
security: []
...
paths:
  /chat:
    post:
      tags: [问答]
      summary: 提交问题并获取回答
      description: 优先检索本地知识库，未命中时调用 DeepSeek，并将回答推送到 SSE 客户端。
      operationId: chatApi
      x-source-location: server/qa_server.py:485
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
            example:
              question: 什么是近视？
              session_id: screen-session-01
      responses:
        '200':
          description: 问答结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
              example:
                user_question: 什么是近视？
                robot_answer: 近视会让远处的东西看起来模糊，需要好好保护眼睛哦！
                source: kb
        '422':
          $ref: '#/components/responses/ValidationError'
...
  /mic/notify_asr:
    post:
      tags: [麦克风]
      summary: 回传 ASR 识别文本
      description: 语音转发节点收到 ASR 结果后回传文本，并推送 `asr_received` 事件。
      operationId: notifyAsr
      x-source-location: server/qa_server.py:593
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AsrNotifyBody'
            example:
              text: 看多久要休息眼睛？
      responses:
        '200':
          description: 回传成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusOkResponse'
        '422':
          $ref: '#/components/responses/ValidationError'
...
components:
  responses:
    ValidationError:
      description: 请求参数校验失败
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/HTTPValidationError'
  schemas:
    ChatRequest:
      type: object
      required: [question]
      properties:
        question:
          type: string
          description: 用户问题；仅包含空白字符时会返回 `source=error` 的 200 响应。
        session_id:
          type: string
          default: default
          description: 会话标识，用于保留最多 10 轮、30 分钟有效的对话历史。
    ChatResponse:
      type: object
      required: [user_question, robot_answer, source]
      properties:
        user_question:
          type: string
        robot_answer:
          type: string
        source:
          type: string
          enum: [kb, deepseek, error, agent]
          description: 回答来源：本地知识库、DeepSeek、错误兜底，或已核验的 Agent 回答。
...
    AsrNotifyBody:
      type: object
      required: [text]
      properties:
        text:
          type: string
          description: ASR 识别出的原始文本。
    StatusOkResponse:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [ok]
...
    HTTPValidationError:
      type: object
      properties:
        detail:
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
    ValidationError:
      type: object
      required: [loc, msg, type]
      properties:
        loc:
          type: array
          items:
            oneOf:
              - type: string
              - type: integer
        msg:
          type: string
        type:
          type: string